requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
asttokens==2.0.5
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
importlib-metadata==7.0.1
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.2
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.7.2
kiwisolver==1.4.7
labelImg==1.8.6
lightning-utilities==0.11.9
lmdb==1.6.2
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==25.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==5.9.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.15.1
PyMySQL==1.1.1
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2025.1
# pywin32==305
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.10.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
stack-data==0.2.0
sympy==1.13.3
timm==1.0.14
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2025.1
tzlocal==5.3
ultralytics==8.3.74
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.5
wheel==0.44.0
yarl==1.15.2
zipp==3.20.2
